Try to load image with Pillow
Instead of testing whether the file is still open. This should help in
waiting for the FTP client on the camera to finish uploading the image.

+++ pumaguard/presets.py
@@ -342,7 +342,7 @@ class Preset:
         self._notebook_number = notebook
 
     @property
-    def file_stabilization_extra_wait(self) -> int:
+    def file_stabilization_extra_wait(self) -> float:
         """
         Get extra wait.
         """
